Cleanups in tree.c.
(Logical change 1.3742)

+++ include/reiser4/tree.h
@@ -65,7 +65,7 @@ extern void reiser4_tree_set_height(reiser4_tree_t *tree,
 extern void reiser4_tree_pack_set(reiser4_tree_t *tree,
 				  pack_func_t func);
 
-extern int32_t reiser4_tree_fetch(reiser4_tree_t *tree,
+extern int64_t reiser4_tree_fetch(reiser4_tree_t *tree,
 				  reiser4_place_t *place,
 				  trans_hint_t *hint);
